added about page, home page and integrated React Router Dom V6 to the project
