+++ src/App.tsx
@@ -1,43 +1,24 @@
 // * Modules
-import { useState } from 'react';
-import FeedbackData from './data/FeedbackData';
-import { v4 as uuidv4 } from 'uuid';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 // * Components
 import Header from './components/Header';
-import FeedbackList from './components/FeedbackList';
-import FeedbackStats from './components/FeedbackStats';
-import FeedbackForm from './components/FeedbackForm';
+import Home from './pages/Home';
+import About from './pages/About';
+import AboutIconLink from './components/AboutIconLink';
 
 export default function App() {
-  // * Application State that contains the feedback data
-  const [feedback, setFeedback] = useState(FeedbackData);
-
-  // * Functionality: Add a new feedback
-  const addFeedback = (newFeedback: {
-    text: string;
-    rating: number;
-    id: any;
-  }) => {
-    newFeedback.id = uuidv4();
-    setFeedback([newFeedback, ...feedback]);
-  };
-
-  // * Functionality: Delete a feedback
-  const deleteFeedback = (id: number) => {
-    if (window.confirm('Are you sure you want to delete this feedback?')) {
-      setFeedback(feedback.filter((item: { id: number }) => item.id !== id));
-    }
-  };
-
   return (
-    <div className='App'>
+    <Router>
       <Header />
       <div className='container'>
-        <FeedbackForm handleAdd={addFeedback} />
-        <FeedbackStats feedback={feedback} />
-        <FeedbackList feedback={feedback} handleDelete={deleteFeedback} />
+        <Routes>
+          <Route path='/' element={<Home />} />
+          <Route path='about' element={<About />} />
+        </Routes>
       </div>
-    </div>
+
+      <AboutIconLink />
+    </Router>
   );
 }
