+++ src/index.scss
@@ -197,6 +197,7 @@ input:focus {
 .about-link a {
   color: #fff;
   cursor: pointer;
+  margin-right: 30px;
 }
 
 .about-link a:hover {
